feat(inventory): 重构物品找回系统和库存序列化机制
- 替换ObjectMapper为JSON-java库处理物品数据序列化
- 添加ItemConverter工具类用于物品对象转换
- 重构物品找回脚本优化性能和代码结构
- 实现Equip类的Lombok注解简化getter/setter方法
- 添加JacksonConfig全局JSON序列化配置
- 优化FredrickProcessor中物品取回逻辑
- 修改DueypackagesDO实体类结构支持单物品包裹
- 实现物品稀疏存储减少数据库空间占用
- 修复物品找回费用计算准确性问题
- 优化装备属性显示和费用计算展示界面
- 统一物品序列化接口支持不同场景需求

+++ gms-server/src/main/java/org/gms/net/server/channel/handlers/ScrollHandler.java
@@ -76,7 +76,7 @@ public final class ScrollHandler extends AbstractPacketHandler {
                 }
 
                 short oldLevel = toScroll.getLevel(); // 记录装备的原始等级
-                byte oldSlots = toScroll.getUpgradeSlots(); // 记录装备的原始升级插槽数量
+                short oldSlots = toScroll.getUpgradeSlots(); // 记录装备的原始升级插槽数量
                 Inventory useInventory = chr.getInventory(InventoryType.USE); // 获取玩家的使用栏库存
                 Item scroll = useInventory.getItem(scrollSlot); // 获取使用的卷轴
                 Item wscroll = null;
